Remove mysensors YAML config support
