+++ homeassistant/components/mysensors/const.py
@@ -11,9 +11,6 @@ ATTR_GATEWAY_ID: Final = "gateway_id"
 
 CONF_BAUD_RATE: Final = "baud_rate"
 CONF_DEVICE: Final = "device"
-CONF_GATEWAYS: Final = "gateways"
-CONF_NODES: Final = "nodes"
-CONF_PERSISTENCE: Final = "persistence"
 CONF_PERSISTENCE_FILE: Final = "persistence_file"
 CONF_RETAIN: Final = "retain"
 CONF_TCP_PORT: Final = "tcp_port"
