+++ tests/components/mysensors/test_config_flow.py
@@ -14,7 +14,6 @@ from homeassistant.components.mysensors.const import (
     CONF_GATEWAY_TYPE_MQTT,
     CONF_GATEWAY_TYPE_SERIAL,
     CONF_GATEWAY_TYPE_TCP,
-    CONF_PERSISTENCE,
     CONF_PERSISTENCE_FILE,
     CONF_RETAIN,
     CONF_TCP_PORT,
@@ -400,332 +399,267 @@ async def test_config_invalid(
 
 
 @pytest.mark.parametrize(
-    "user_input",
-    [
-        {
-            CONF_DEVICE: "COM5",
-            CONF_BAUD_RATE: 57600,
-            CONF_TCP_PORT: 5003,
-            CONF_RETAIN: True,
-            CONF_VERSION: "2.3",
-            CONF_PERSISTENCE_FILE: "bla.json",
-        },
-        {
-            CONF_DEVICE: "COM5",
-            CONF_PERSISTENCE_FILE: "bla.json",
-            CONF_BAUD_RATE: 57600,
-            CONF_TCP_PORT: 5003,
-            CONF_VERSION: "2.3",
-            CONF_PERSISTENCE: False,
-            CONF_RETAIN: True,
-        },
-        {
-            CONF_DEVICE: "mqtt",
-            CONF_BAUD_RATE: 115200,
-            CONF_TCP_PORT: 5003,
-            CONF_TOPIC_IN_PREFIX: "intopic",
-            CONF_TOPIC_OUT_PREFIX: "outtopic",
-            CONF_VERSION: "2.4",
-            CONF_PERSISTENCE: False,
-            CONF_RETAIN: False,
-        },
-        {
-            CONF_DEVICE: "127.0.0.1",
-            CONF_PERSISTENCE_FILE: "blub.pickle",
-            CONF_BAUD_RATE: 115200,
-            CONF_TCP_PORT: 343,
-            CONF_VERSION: "2.4",
-            CONF_PERSISTENCE: False,
-            CONF_RETAIN: False,
-        },
-    ],
-)
-async def test_import(hass: HomeAssistant, mqtt: None, user_input: dict) -> None:
-    """Test importing a gateway."""
-
-    with patch("sys.platform", "win32"), patch(
-        "homeassistant.components.mysensors.config_flow.try_connect", return_value=True
-    ), patch(
-        "homeassistant.components.mysensors.async_setup_entry",
-        return_value=True,
-    ):
-        result = await hass.config_entries.flow.async_init(
-            DOMAIN, data=user_input, context={"source": config_entries.SOURCE_IMPORT}
-        )
-        await hass.async_block_till_done()
-
-    assert result["type"] == "create_entry"
-
-
-@pytest.mark.parametrize(
     "first_input, second_input, expected_result",
     [
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "same2",
             },
             {
-                CONF_DEVICE: "mqtt",
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "same2",
             },
-            (CONF_TOPIC_IN_PREFIX, "duplicate_topic"),
+            FlowResult(type="form", errors={CONF_TOPIC_IN_PREFIX: "duplicate_topic"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "different1",
                 CONF_TOPIC_OUT_PREFIX: "different2",
             },
             {
-                CONF_DEVICE: "mqtt",
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "different3",
                 CONF_TOPIC_OUT_PREFIX: "different4",
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "different2",
             },
             {
-                CONF_DEVICE: "mqtt",
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "different4",
             },
-            (CONF_TOPIC_IN_PREFIX, "duplicate_topic"),
+            FlowResult(type="form", errors={CONF_TOPIC_IN_PREFIX: "duplicate_topic"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "different2",
             },
             {
-                CONF_DEVICE: "mqtt",
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "different1",
                 CONF_TOPIC_OUT_PREFIX: "same1",
             },
-            (CONF_TOPIC_OUT_PREFIX, "duplicate_topic"),
+            FlowResult(type="form", errors={CONF_TOPIC_OUT_PREFIX: "duplicate_topic"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "different2",
             },
             {
-                CONF_DEVICE: "mqtt",
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_VERSION: "2.3",
                 CONF_TOPIC_IN_PREFIX: "same1",
                 CONF_TOPIC_OUT_PREFIX: "different1",
             },
-            (CONF_TOPIC_IN_PREFIX, "duplicate_topic"),
+            FlowResult(type="form", errors={CONF_TOPIC_IN_PREFIX: "duplicate_topic"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "127.0.0.1",
                 CONF_PERSISTENCE_FILE: "same.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "same.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            ("persistence_file", "duplicate_persistence_file"),
+            FlowResult(
+                type="form", errors={"persistence_file": "duplicate_persistence_file"}
+            ),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "127.0.0.1",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "same.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "127.0.0.1",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "different1.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "different2.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            ("base", "already_configured"),
+            FlowResult(type="form", errors={"base": "already_configured"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "different1.json",
                 CONF_TCP_PORT: 343,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_PERSISTENCE_FILE: "different2.json",
                 CONF_TCP_PORT: 5003,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.2",
                 CONF_TCP_PORT: 5003,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_TCP,
                 CONF_DEVICE: "192.168.1.3",
                 CONF_TCP_PORT: 5003,
                 CONF_VERSION: "2.3",
-                CONF_PERSISTENCE: False,
-                CONF_RETAIN: False,
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
-                CONF_TCP_PORT: 5003,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "different1.json",
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
-                CONF_TCP_PORT: 5003,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "different2.json",
             },
-            ("base", "already_configured"),
+            FlowResult(type="form", errors={"base": "already_configured"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM6",
                 CONF_BAUD_RATE: 57600,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
-                CONF_TCP_PORT: 5003,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
                 CONF_BAUD_RATE: 115200,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "different1.json",
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
                 CONF_BAUD_RATE: 57600,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "different2.json",
             },
-            ("base", "already_configured"),
+            FlowResult(type="form", errors={"base": "already_configured"}),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM5",
                 CONF_BAUD_RATE: 115200,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "same.json",
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM6",
                 CONF_BAUD_RATE: 57600,
-                CONF_RETAIN: True,
                 CONF_VERSION: "2.3",
                 CONF_PERSISTENCE_FILE: "same.json",
             },
-            ("persistence_file", "duplicate_persistence_file"),
+            FlowResult(
+                type="form", errors={"persistence_file": "duplicate_persistence_file"}
+            ),
         ),
         (
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_MQTT,
                 CONF_DEVICE: "mqtt",
                 CONF_PERSISTENCE_FILE: "bla.json",
-                CONF_BAUD_RATE: 115200,
-                CONF_TCP_PORT: 5003,
                 CONF_VERSION: "1.4",
             },
             {
+                CONF_GATEWAY_TYPE: CONF_GATEWAY_TYPE_SERIAL,
                 CONF_DEVICE: "COM6",
                 CONF_PERSISTENCE_FILE: "bla2.json",
                 CONF_BAUD_RATE: 115200,
-                CONF_TCP_PORT: 5003,
                 CONF_VERSION: "1.4",
             },
-            None,
+            FlowResult(type="create_entry"),
         ),
     ],
 )
@@ -734,7 +668,7 @@ async def test_duplicate(
     mqtt: None,
     first_input: dict,
     second_input: dict,
-    expected_result: tuple[str, str] | None,
+    expected_result: FlowResult,
 ) -> None:
     """Test duplicate detection."""
 
@@ -746,12 +680,17 @@ async def test_duplicate(
     ):
         MockConfigEntry(domain=DOMAIN, data=first_input).add_to_hass(hass)
 
+        second_gateway_type = second_input.pop(CONF_GATEWAY_TYPE)
         result = await hass.config_entries.flow.async_init(
-            DOMAIN, data=second_input, context={"source": config_entries.SOURCE_IMPORT}
+            DOMAIN,
+            data={CONF_GATEWAY_TYPE: second_gateway_type},
+            context={"source": config_entries.SOURCE_USER},
+        )
+        result = await hass.config_entries.flow.async_configure(
+            result["flow_id"],
+            second_input,
         )
         await hass.async_block_till_done()
-        if expected_result is None:
-            assert result["type"] == "create_entry"
-        else:
-            assert result["type"] == "abort"
-            assert result["reason"] == expected_result[1]
+
+        for key, val in expected_result.items():
+            assert result[key] == val  # type: ignore[literal-required]
